+++ app/src/main/res/layout/activity_search_place.xml
@@ -4,7 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-
+        <variable
+            name="viewModel"
+            type="com.dnd_8th_4_android.wery.presentation.ui.write.place.viewmodel.SearchPlaceViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -99,6 +101,8 @@
                 android:hint="@string/search_place_hint"
                 android:paddingHorizontal="26dp"
                 android:paddingVertical="16dp"
+                android:inputType="text"
+                android:imeOptions="actionSearch"
                 android:textColor="@color/black"
                 android:textColorHint="@color/gray500"
                 app:layout_constraintBottom_toBottomOf="parent"
@@ -118,6 +122,7 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
         <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rv_search_result"
             android:layout_width="0dp"
             android:layout_height="0dp"
             android:layout_marginVertical="16dp"
